qualify source_name if the uri is a sub directory of a git repo (#294)
Toward #275.

Follow up of #276.

Based on the suggestion by @mateiz, this PR qualifies the source_name of a project run when the uri is actually a sub directory of a git repo.

With `mlflow run mlflow/example/tutorial -P alpha=0.4` run, the UI will now show 

![image](https://user-images.githubusercontent.com/10642337/44133100-9ecc66aa-a012-11e8-8d98-20d848d49393.png)

in `Run Command`. The source name now is a git repo path. And the git commit version is also logged in the command.

Please have a look.