Make compact view default (#700)
Originally, we made the compact view non-default because it didn't support column-splitting or persistence -- now that we have both of these, we should go ahead and make this view default. Users can still switch back to the old view, and that choice will be persisted.

Also switched ordering of the icons so the default is on the left.