Added Bash command for UI
In section "Comparing the Models" the command to launch the MLflow UI was given for R mlflow_ui(), but not for Bash  mlflow ui. This took me a bit to figure out since I needed the Bash command because I'm working in Python.  

I copied the relevant part from the Quickstart page to include commands for both, as well as the link to the page to view the results.