Pin sklearn version in test dependencies (#731)
The latest version of scikit-learn (0.21.0) is only available on pip. It has not yet been released on Anaconda. This causes test failures when attempting to resolve default conda environments: conda attempts to install a version of the scikit-learn dependency that is not available.

To resolve this issue, we will temporarily pin the version of scikit-learn to the latest available version on Anaconda (0.20.0).