Wrap the RunInfo version in an abbr tag
When using experiment view, the version extracted from runInfo is truncated to 6 characters. In cases where an alternative version number is supplied that is not a git hash (e.g. semantic version) , it is useful to still have a way of seeing the entire version number in the experiment view without cluttering the interface.