Update tests to new expected behaviour
This change updates the fluent tests to expect that the source_name and
source_type arguments to start_run are always respected, even in the
case of being in a databricks notebook.

The pre-existing behaviour was that these arguments were always ignored
when in a databricks notebook. Preserving this logic in the new context
provider system would have been awkward, and it probably makes sense to
not ignore these arguments anyway.