Remove values for project env tag from `mlflow.utils.mlflow_tags`
Currently, all other values in `mlflow.utils.mlflow_tags` are tag keys,
and there is nothing to indicate in that module that `MLFLOW_DOCKER` and
`MLFLOW_CONDA` are actually intended for use as tag values.

At some point, an enumeration of valid values for this tag could be
created, but since that seems outside the scope of this PR and these tag
values are only used in a single location, I've just baked them in place
for the time being.