Add a button for downloading artifacts (#967)
This makes it possible to download .html files, images, etc and view them locally. It looks like this:

![screen shot 2019-03-07 at 7 24 25 pm](https://user-images.githubusercontent.com/228859/54005533-ae3b9a00-410e-11e9-92c8-451e2224d63c.png)

I also updated `package.json` to have it treat /get-artifact as a web server, because otherwise it tried to redirect all paths that return MIME type HTML to /index.html (this is the default behavior for the React proxy).