Provide default implementations of AbstractStore log methods
The abstract store has methods for logging individual metrics, params
and tags, however since these are not marked as abstract methods they're
not currently required to be implemented in tracking store
implementations.

This change provides default implementations of `log_metric`,
`log_param` and `set_tag` that proxy calls to the more general
`log_batch` added by @smurching in #950.

Following this change, as long as stores implement `log_batch`, the
above methods need not be provided.