rest-api.rst updates (#1033)
The current rest-api.rst file in the repo seems out of sync with the published docs (https://www.mlflow.org/docs/latest/rest-api.html), so trying to synchronize them. The file currently is generated from PB + some manually added information (see #926). We should try and make it consistent with the PB without having to manually edit it, but for now putting it in a workable state.