Fix RunsArtifactRepository.list_artifacts (#1175)
Have RunsArtifactRepository.list_artifacts actually return the artifacts.

This was found as I was testing the changes in #1174. We should really have unit/integration tests specifically for RunsArtifactRepository but skipping for now since the codepath will be covered by other modules (e.g. the sagemaker tests mentioned above) and I'd like to get this in quickly to unblock the API changes to come.