Prevent bad experiment names from rendering the application unusable
Point MissingConfigException to its parent class: MlflowException.

Issue https://github.com/mlflow/mlflow/issues/1461 explains how a badly
named directory in `mlruns` can cause an MlflowException to be raised
when listing experiments.

The desired behavior when a misnamed directory is met is to log the
error, and continue with the good directories, not to fail without
warning.

mlflow does not log MlflowException and only inspecting browser logs
reveals errors. mlflow does catch a subclass MissingConfigException.

This commit makes MissingConfigException a reference to its parent
class, as it is not such an important case that warrants its custom
exception.