Docker projects should preserve directory structure (#1436)
Fixed issue where directory structure for a docker based MLproject was not preserved in the final docker image. If an entry point wasn't at the top level of the project, the path to call it in the commands section MLproject file could not match the actual location in the project itself. This could be worked around but made any complicated project development not feasible.