Fix OSS docs build bug (#1607)
Fixes bug where running make html to build the OSS docs would result in the removal of Javadocs by the R doc build, which deletes the docs/build directory so that Sphinx docs aren't mistakenly included in the R docs. The fix for now is to run the R build first.