Deal with race condition where multiple runs upload projects files to dbfs (#1641)
If you are running mlflow run with the same project in parallel with --backend databricks, it's possible that multiple runs will try to upload the projects file onto dbfs (because even though mlflow checks whether the file exists before uploading, at the start of the upload the file does not exist). Here we handle the "File already exists" error that happens with dbfs in such cases.