Showing a loading message in Artifacts section of Run page while resu… (#18)
* Showing a loading message in Artifacts section of Run page while results are fetched.

In order to support this behaviour, the following changes were made:

* RequestStateWrapper now is aware of a list of requests for which
it should not wait before calling the `children` function. This
list of requests are called `asyncRequests` and are as well passed
when calling the `children` function.

* In the render logic of `RunPage.js` we can query the state of the
request in charge of loading the artifacts and we propagate that information
through `RunView.js` -> `ArtifactPage.js` -> `ArtifactPage.js` -> `ArtifactView.js`
so that a nice loading animation is shown when appropiate.