In fluent API, add note about active run instantiation side effect
In some APIs such as `get_artifact_uri`, the function starts an active
run if no there is no active run. This should be considered a
side-effect because these it is not apparent that these functions would
instantiate the active run.

As a short-term solution, I have updated the documentation in
functions that have this behavior.