Refresh experiment list after deletion
Updated experiment deletion functionality such that
the shown experiment list is refreshed/updated after
performing the deletion.
To do so, the behavior of experimentById reducer had
to be modified.
The previous implementation kept deleted experiments in the state
until the page was refreshed. This could also be observed when
deleting an experiment via the CLI since
the deleted experiment would remain in the UI until refreshing.