Manage Experiments from MLflow UI (Create/Rename/Delete Experiments) (#2348)
* Add delete and update experiment endpoints to service

Added delete and update experiment endpoints to existing
MlflowService implementation. These operations were already supported by
MLflows REST API. For more information, see:
https://www.mlflow.org/docs/latest/rest-api.html
Added corresponding AJAX calls.

* Add edit and delete button to experiment list

Added edit and delete button to experiment list view.
The buttons do not have any functionality associated with them yet.

* Add delete experiment functionality

Added functionality to delete experiments.
When clicking on the delete button in the experiment list,
 confirmation modal is opened.

* Refactor and generalize modal to rename runs

Updated modal to rename runs such that it is more generic.
In particular, updated the displayed form.
As such, this component can be used for renaming experiments.

* Update delete experiment modal

Updated delete experiment modal such that the passed experimentId is
a number instead of a string. Updated ExperimentListView accordingly.

* Add rename experiment functionality

Added functionality to rename experiments.
When clicking on the edit button in the experiment list,
 a modal is opened asking for a new experiment name.

* Fix rename experiment functionality

Fixed a minor issue which caused the experiment name to not be updated.

* Refresh experiment list after deletion

Updated experiment deletion functionality such that
the shown experiment list is refreshed/updated after
performing the deletion.
To do so, the behavior of experimentById reducer had
to be modified.
The previous implementation kept deleted experiments in the state
until the page was refreshed. This could also be observed when
deleting an experiment via the CLI since
the deleted experiment would remain in the UI until refreshing.

* Fix existing unit tests

* Add unit test

Added unit tests testing that the delete and rename
experiment modals are opened after clicking the
respective links in the experiment list.

* Update styles of rename experiment modal

* Disable delete experiment link for active experiment

Updated the experiment list in such a way that
the "Delete Experiment" option is disabled
for the currently active/selected experiment.

* Fix linting issue

* Migrate renaming modals to antd

Based on the following input of the created PR:
https://github.com/mlflow/mlflow/pull/2348#issuecomment-579087684
Updated rename modals to use antd modals and forms instead of ReactModal and Formik.

* Add focus and initialValue to input in rename modals

Updated RenameFormView used in rename modals.
The input of the modal will be focused and selected after (re)opening the modal.
The input of the modal will contain the respective name of the experiment/run as the initial value
after re(opening) the modal.

* Create generic input modal & refactor rename modals

Refactored rename modals.
Since the RenameExperimentModal and RenameRunModal had a lot in common,
I put most code in a generic modal component.
This generic modal component is used by the rename modals.

* Migrate ConfirmModal to antd

Based on the following input of the created PR:
https://github.com/mlflow/mlflow/pull/2348#issuecomment-579087684
Updated the generic ConfirmModal to use antd modals.

* Update "Delete" icon

* Fix error handling of generic modal

Fixed error handling of generic modal.
Before this change, the error modal was not correctly displayed.

* Align with naming conventions & use wrapped dispatch functions explicitly

Incorporated feedback from PR.
Aligned with naming conventions (callbacks, variables...)
Moved follow-up calls from deleteExperimentApi and deleteRunApi up to be explicitly invoked
as a follow up in the then blocks.
Updated code to use wrapped dispatch functions explicitly.

* Add basic rendering test for InputFormView

* Update GenericInputModal

Refactored GenericInputModal, making the component more generic.
1. A form component can now be passed to the modal as a render prop
2. The entire values object (from the form) is now passed to the submit callback.
The caller can thus decide what to do with the values object.

* Align with naming conventions inside basic rendering test

* Rename InputFormView to RenameForm

* Allow users to delete the current experiment

Updated the experiment deletion behavior.
Users can now delete the currently selected experiment.
As a consequence, they are re-routed to the root which then picks the next active experiment to show.

* Add "Create Experiment" functionality

Added functionality to create experiments from the UI.
Added "Create Experiment" button based on mockup in PR #2348.
Added modal and form that handles experiment creation.
Added basic unit tests.

* Route user to new experiment page after creation

Updated the behavior when creating an experiment.
After creation, the user will be redirected to the newly created experiment page.
As such, they don not need to manually find the new experiment in the list

* Add getExperimentByName endpoint to service

Added getExperimentByName to MLflowService.
https://www.mlflow.org/docs/latest/rest-api.html#get-experiment-by-name

* Update error handling & validation checks

Unified modal error handling based on the discussion in PR #2348.
Updated Rename/Create experiment modals to check whether the experiment name is already in use.
Leveraged async-validator to do the validation against the
current experiment name in the form input.
Added debouncing to the validation trigger to prevent excessive calls.
Other errors, e.g. Internal Server Errors, are now displayed in a message instead of a error dialog.

* Update debounced validator usage & use explicit exports

Updated debounced validator. Before, a new instance of the validator function was created per render.
Bound debounce outside the render.
Updated some components to use explicit exports.

* Add unit test for experimentNameValidator

* Add artifact location to CreateExperimentModal

Updated CreateExperimentModal. A user can now specify an artifact location.
Creating an experiment now matches the overall create_experiment signature

* Update named exports & update lodash import

* Update updateExperimentApi

Moved follow-up getExperimentApi call to the rename modal layer.

* Fix NEP error after experiment creation

Updated the experiment creation behavior.
Before, the sequence of the performed calls after a user created an experiment was not guaranteed.
As such, the experiment list was sometimes not refreshed before the rerouting happened which led to an error in the UI.

* Update formatting

* Update UUID handling in CreateExperimentModal

* Update experiment creation to use async/await pattern