[SERVER UI] FTP Parsing Workaround (#2196)
* FTP Parsing Workaround

The ftp.nlst() function at line 106 creates a list that starts with "." and "..". These two items are not objects so, when accessing through FTP, MLflow throws an error about not being able to access those files. In this workaround, we iterate the list starting at index 2.

* Change FTP Workaround Method

Instead of parsing the list from index 2 the list is filtered beforehand to remove "." and "..".

* Wrong Variable Name

Co-authored-by: Mani Parkhe <mani@databricks.com>