Onnx.predict handles list response
MLflow assumes that an onnx model will return a numpy array. A user
discovered this is not the case and in fact, the model can return a
python list.

The result is that the service throws an error when it attempts to call
reshape() on a python list and that function does not exist.

The fix is to look onvert a list to numpy array before calling reshaping it.