Pin cloudpickle to 1.4.1
Previously, MLflow expected users to stay with one version of
cloudpickle when loading and saving models within their conda
environment.

Feedback from the community has shown that this is not common. Many
users will upgrade to the latest version available. We originally tried
to allievate this problem by throwing a warning that the cloudpickle
versions between the save and load were different but that has proven
still unfavorable.

This PR is intended to further address this user pain by pinning a
default cloudpickle version to 1.4.1. Users should be free to select
their own cloudpickle version but with a default, this issue will be
less of a footgun for most users.