Handle spaces in Databricks profile URIs (#3338)
Do not allow spaces in the key prefix of Databricks profile URIs. We could strip the tokens ourselves, but since there are multiple branches in the parsing code, it's prone to error. It seems reasonable to expect the user to send in a clean version of the URI (without leading or trailing spaces).