Fix a bug that converts int64 to string when converting Protobuf to JSON (#5010)
* Fix an issue with the timestamps in the endpoint response

Signed-off-by: Chenran Li <chenran.li@databricks.com>

According to issue #4037, the returned JSON of the endpoints has `creation_timestamp` and `last_updated_timestamp` as strings, not numbers. It's different from what was documented in the [official doc](https://www.mlflow.org/docs/latest/rest-api.html#mlflowregisteredmodel).

The reason is we are calling Google's `MessageToJson` API to convert protobuf to json, which implicitly converts int64/fixed64/unit64 fields to strings. And they claimed it's a feature not a bug (see the [discussion](https://github.com/protocolbuffers/protobuf/issues/2954#issuecomment-292671661)).

According to the bug reporter, this bug doesn't exist in Azure ML mlflow server (which is essentially our Databricks mlflow server). That's because we are using ScalaPB's `ToJson()` API for all the Databricks endpoints, and it doesn't convert int64 to string.

There is no way to let `MessageToJson` API not convert int64 to strings. Nor are there any other good Python proto-to-json libraries. So to fix this bug, we have to choose from:
* (what I'm doing in this PR) manually converting the int64/uint64/fixed64 fields back to numbers after calling `MessageToJson`
* (too risky so I chose not to do) writing our customized `MessageToJson` API