Changed DefaultAzureCredential to ChainedTokenCredential for azure client
Before the Az client used the DefaultAzureCredential for auth. This looks
through a few diffent settings for auth. Unfortionately, something with
one of the steps is not working and it gets stuck...

So, changed to ChainedTokenCredential, where is is possible to set
the steps that will be checked. Now, auth will be handled with:
1. Environment variables with EnvironmentCredential
2. Azure CLI with AzureCliCredential

Signed-off-by: Herman Johnsson herman.jonsson92@gmail.com