Improve run compare ui (#5310)
Run compare UI improvements:

Move the plots to the top of the page, the comparison table to the bottom.
Add titles to the plots: Visualizations, comparison table: Run details. And make them collapsible.
Reduce the default number of selected parameters from 10 to 3.
In the hyperparameter tuning examples (search_random.py and search_hyperopt.py), log parameters for each trial run, to make it easier to reproduce the trials that are not the best, and easier to generate a parallel coordinates plot.