Smarter deletion of python modules in sys.modules
If you organise your code in modules then the current deletion of the modules in sys.models will not work. 
Let's say you have a models folder in you your source folder with python files and a '__init__.py' inside. The path of a python file would look like 'src/models/model_1.py'.
The current method tries to delete 'modules' in sys.modules which does not exist.
What we actually want to delete is ['src, 'src.models', 'src.models.model_1']. This is now implemented on line 130-131.

Signed-off-by: michaguenther <58656683+michaguenther@users.noreply.github.com>