Adding inline examples of model flavor usage with MLflow for fastai.
This commit addresses the fastai portion of feature request #6099.

I added inline code examples that show how to use automatic
logging, save a fastai model, and load it in native fastai
format and in pyfunc format.  These examples are based on
the fastai tabular example and examples from the mlflow.fastai
module.

Signed-off-by: Randy DeFauw <rddefauw@amazon.com>