Fixes #6028: Don't add /code subdirectories to sys.path
MLflow was not only adding `<model>/code` to the PYTHONPATH, but also
all of the directories underneath `<model>/code` which caused unexpected
import conflict errors with built-ins and 3rd-party libraries.

This change makes it so that only `<model>/code` is placed on the system
path, as users are expected to import from the top-level directories
directly themselves.

More details, including minimal reproducible example, can be found in
issue \#6028

Signed-off-by: John Allard <john@jhallard.com>