Dockerfile optimization
The command to pull the setup_16.x script was added due to the fact that there is already apt update command inside the script, executing this command a second time had no effect, it only increased the image build time, the curl and installation was also removed due to the fact that the image is python3.8 already has curl and default-libmysqlclient-dev, apt jarn and npm cache removal has also been added. There was also no need to separate the installation into two layers, so they were combined into one layer.

Signed-off-by: alekseyolg <86011874+alekseyolg@users.noreply.github.com>