Hide experiments whose name starts with a dot
Filter the default list experiment on the web frontend and hide experiments
starting with a dot. As soon as the search box is not empty, any matching
experiment is visible regardless of a preceding dot.

I've implemented this feature in an on-premise deployment to hide experiments
solely used by automated MLOps end-to-end tests. I think it is nice to have an
easy way to hide experiments that don't need manual inspection (most of the
time) and would otherwise distract from the actual experiments.

Signed-off-by: Frank Sauerburger <frank@sauerburger.com>