Fix _PyTorchWrapper predict when cuda device
If you select device='cuda', this line will give an error as it is trying to convert a cuda tensor to numpy. It needs to be sent to cpu first. Note that this code will work with device='cpu' as the tensor will be already in cpu, and no action will be taken.

Signed-off-by: Hector Anadon <hectoranadonleon@gmail.com>