gateway: handle max retries exceeded exception
This PR updates the http response handling logic in gateway. In case of
max retries exceeded, this exception is catched and one more request is
made to the server without retries. In this way, the actual response can
be fetched and displayed to the user.

When max_retries is set to 0, do not set any retries config for HTTPAdapter
and use the default config (no retries). An unit test is added for this.

Signed-off-by: Shaotong Li <shaotong.li@databricks.com>