stripped the extra "/" from dest_path
The destination path variable `dest_path` in both functions `list_artifacts()` & `delete_artifacts()` had an extra slash. This extra "/" prevented the `mlflw gc` command from deleting artifacts when MinIO is used.

This happens because MLflow tries to list artifacts using a prefix with two trailing slashes. This is considered to be an invalid request by MinIO since it is built upon a filesystem-based storage backend which cannot use the "/" character for anything but the file path separator.

Stripping the extra slash "/" should solve the problem.

Signed-off-by: Tarek Ashraf <tashraf154@gmail.com>