Update docstrings for log_model and load_model functions in mlflow.diviner module.
Enhanced and clarified the docstrings for the log_model and load_model functions in the Diviner module.
Added detailed descriptions for each parameter, including examples.
Provided explanations for optional configuration options and included code examples for better understanding.
Updated the documentation to make it easier for users to utilize these functions for logging and loading models.

Signed-off-by: Pasindu Mahima Munasinghe <pasindu.mahima@gmail.com>