Store failed Python tests as GitHub Action summary
This change adds functionality to write failed test information to the
GitHub Actions job summary when running in CI. When the GITHUB_STEP_SUMMARY
environment variable is set (which is automatically set in GitHub Actions),
the failed test command will be appended to the job summary for easy access.

This makes it easier to quickly see and re-run failed tests directly from
the GitHub Actions UI without having to dig through the full test logs.

Signed-off-by: Harutaka Kawamura <harutaka.kawamura@databricks.com>
Signed-off-by: harupy <17039389+harupy@users.noreply.github.com>