Fix trace status mapping: UNSET span status should map to IN_PROGRESS
STATE_UNSPECIFIED should never be used for a trace that has logged spans.
When a span has UNSET status (the default OpenTelemetry status), the trace
should be IN_PROGRESS, not STATE_UNSPECIFIED. This makes much more sense
semantically - if we have spans but their status is unset, the trace is
still in progress.

Signed-off-by: dbczumar <corey.zumar@databricks.com>