Fix missing artifact location tag when creating trace via log_spans
When log_spans creates a new trace in SqlAlchemyStore, it was missing
the artifact location tag that's required for search_traces to work.
This caused traces created via the OTel endpoint to not be retrievable.

Now we add the artifact location tag when creating a trace, matching
the behavior of start_trace and deprecated_start_trace_v2 methods.

Signed-off-by: dbczumar <corey.zumar@databricks.com>