Mock server version check only when MLflow version < 3.4
Add conditional mocking of the server version check in the OTLP endpoint test
to allow the test to pass on older MLflow versions that don't have the OTLP
endpoint. The mock is only applied when the current MLflow version is less
than 3.4, ensuring we test the actual version check on newer versions.

Signed-off-by: dbczumar <corey.zumar@databricks.com>