Fix substitution of escape sequences in genai format_prompt
In mlflow.genai.judges.utils.format_prompt, a string is used as the
replacement argument for re.sub. This leads to escape sequences being
interpreted which seems to be undesirable (for example, when the
function is used to judge retrieval sufficiency, a literal `\n`
(backslash followed by n) in the retrieval context would be replaced by
a newline). Also, unknown escape sequences such as `\x` or `\u` lead to
judge errors since re.sub throws an exception when they are encountered.

Fix this by using the function-call form for the replacement argument
which uses the string literally.

Signed-off-by: Stefan Walk <swalk@ptc.com>