Add workspace database schema
This adds the required workspace columns and the workspace catalog table
with the default workspace precreated. All workspace columns default to
"default" for now and we may choose to remove the defaults once the
tracking store and model registry store are made workspace aware to
catch application logic issues not properly setting the workspace.

Some model registry store changes were needed to account for the new
composite foreign key.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>